bot/requirements.txt
aiogram==3.17.0
httpx==0.27.0
pydantic==2.10.6
python-dotenv==1.0.1
